gethostbyname inet_pton fix
when gethostbyname is provided an IP address in the form of a string, it
is supposed to write the address to h_name; the function instead was
writing an empty string. getaddrinfo does the correct thing by calling
inet_pton and returning on success, and the mistake is in gethostbyname.

The fix is to strlen(result->ai_canonname): If the inet_pton call in
getaddrinfo is successful, ai_canonname will point to valid memory
(non-NULL) but the string will be empty. in this case, the name (which
is an IP address string) will now correctly be copied to h_name of the
hostent structure.