Added getprotobyname/getprotobynumber
The implementations of the getproto* functions follow from the getserv*
functions: same static name allocation, same type of internal function
that opens a file to search, aliases are not written to the struct, same
type of error handling/returns.

Note that in Windows, the protocols file was "<wherever>\\etc\\protocol"
ie the name is in singular.