getaddrinfo AI_PASSIVE memory error
when getaddrinfo is passed name = NULL and AI_PASSIVE in hints->ai_flags, it was
setting the s_addr value to INADDR_ANY but *not* returning the addrinfo
pointer via *res = ai.

This caused a free(NULL) memory error when the caller tried to free res,
because the caller expects res to be a valid pointer to an struct
addrinfo.

Fix was to assign *res = ai before returning in the AI_PASSIVE case.