avoid malloc/free wrt LookupServicesByX
the getservbyX functions allow to specify the protocol as NULL, in which
case services with either protocol can be matched. The previous solution
this involved sending a char **localproto to LookupServicesByX,
allocating *localproto at the callee if it was NULL, and then have the
caller free it if the allocation did take place.

This gets ugly on the caller side, especially with error handling added
to the mix: for example, getnameinfo requires a LookupServicesByPort
call, which requires error-checking, null-checking, and freeing what was
basically a temporary string anyway.

Now the LookupServicesByX functions expect a non-NULL buffer + non-zero
length for localproto, and will write the protocol to the buffer if it
is an empty string. This makes it simpler to handle from the caller
side (can now pass constant strings as well).

Also, you can pass NULL for the official name in LookupServicesByName,
since some callers will want only the port value.