Fix redbean's Lua-function Log, so it can be used in the global scope.
Before this fix Log crashed the redbean when Log was used in the global scope.
The demo versions of redbean has a Log in a main function, therefore it didn't crash.
Lua doesn't need a main function and every other redbean Lua-function can also be used in the most outer scope.