load C extensions of external packages correctly
At present, all C extension modules loaded in the APE Python are
built-in modules -- they are compiled into the interpreter instead of
being separate shared objects in some folder.

In the Python3.6 stdlib, the modules usually follow a naming convention
like: X imports a module called _X, which is built from Modules/_X.c.
(example: json imports _json, which is built from Modules/_json.c).

However, external Python packages follow a different naming convention:
X imports a module called ._Y, which is available as a shared object
located in the same directory as X.

(example: markupsafe imports from ._speedups, expecting a _speedups.so
to be available within the markupsafe folder).

This change makes the interpreter load the required module as a builtin
(if present) if the requisite shared object/python file is not found.
The benefit of this is that the external packages can keep the same
naming convention for their C extensions, and once built, the APE will
load those extensions without error.