shave off stat syscalls by rearranging imports
the hardest import method format to speed up is the "from x import y",
because I have no idea whether y is a module or not.

If y is a module, "from x import y" behaves similarly to the format
"import x.y; y = x.y", which means I can bypass some checks by doing the
import and equality separately.

If y is a function, "from x import y" is not equivalent to the above
format, and the import will trigger a ModuleNotFoundError. Unfortunately
I can't check for or propagate such errors during the default APE
startup without adding a whole bunch of checks, so I'd prefer to avoid
these kinds of imports unless absolutely necessary.

in this PR I avoid three such function imports in io.py and ntpath.py.
The fix is to import the module wholesale (which happens underneath
anyway) and then just check for the attribute - "import x; y = x.y",
essentially similar to how it would be if y was a module.