ttyraster: reset ANSI attributes after each line
This fixes a bug that appears when using `printimage` in my `xterm` or
`gnome-terminal`. When printing a line causes the viewport to scroll,
the last color painted would extend to the last column of the window
rather than the last column of the image. (This is especially noticeable
when rendering a narrow image in a wide window.) A similar artifact
would occur with the last line, which would appear in the color of the
bottom-right corner of the image.

By slapping an `\e[0m` before each newline and before the end of the
image, we avoid these artifacts.

![Screenshots of the output of `printimage` before and after this
change. Each screenshot shows `lemurs.png` printed twice; the first copy
is only partially visible because the window has scrolled. In the
"before" picture, the second copy has extra bands of color off to the
right of the image starting about halfway down, plus a gray band beneath
the image. The "after" picture has no such artifacts.][ss]

[ss]: https://user-images.githubusercontent.com/4317806/178096633-9a9e46a7-360b-49c1-8569-e6b401b81d09.png

wchargin-branch: ttyraster-reset-each-line
wchargin-source: dfd8233b06783f16cf5fd40285f7d1bc022c9cb2