converted json.org tests into lua files
DecodeJson works on strings, so I pasted every JSON file as a
multiline string into a lua file. This required one small change to the
test cases, which is putting a space after every [ or ], so that Lua
doesn't get confused by [[ or ]] which are the multiline string
terminators (see pass2.lua for an example).