throw parsing errors for missing commas and colons
context will be set as zero if a comma is found AFTER_VALUE or a colon
is found before OBJECT_VAL, so we add an UNLIKELY check before parsing
each value to confirm that these two fields are set to zero in context.

this change now means that ljson complies with all the tests from
json.org, and only deviates from JSONTestSuite's expected behavior when
it comes to parsing numbers.