[metal] Place GDT in read/write segment
The CPU absolutely needs to alter the GDT when loading the
task register (via ltr).  To account for this, I move the
GDT into a read/write data section.  There is still a "rump"
read-only GDT in the text section that is used by the real
mode bootloader.

I also delay the loading of the task register (ltr) until
after the IDT and TSS are finally set up.