Fix inttypes.h FAST16 macros to have a correct definition
The PRI and SCN are described by the C standard as being:
> suitable for use within the format argument of a formatted
> input/output function when converting the corresponding integer type
- C Standard, 7.8.1. Macros for format specifiers

However, cosmopolitan fails to do this properly for the FAST16 macros,
meaning that for example, the following program:

 #include <stdio.h>
 #include <stdint.h>
 #include <inttypes.h>
 #undef NDEBUG
 #include <assert.h>

int main()
{
    uint_fast16_t v = 0xFFFFFFFF;
    assert(sscanf("12345", "%" SCNdFAST16, &v) == 1);

    printf("%ju\n", (uintmax_t)v);
}

will print 4294914105 instead of the correct 12345 on
cosmopolitan. This is caused by the fact that the definitions of the
FAST16 macros have been made with the assumption that these are 16-bit
types, when they are in fact defined as 32-bit types.

This patch fixes this.